# api/openapi.yaml
openapi: 3.1.1
info:
  title: DCC Control Server API
  version: 1.3.0
  description: >-
    Contract for native DCC clients. User administration is deliberately absent
    from this public API and is available only over the local administration socket.
servers:
  - url: https://dcc-server.example:8443
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      security: []
      operationId: health
      responses:
        '200':
          description: Server is ready.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { const: ok }
  /api/v1/system/info:
    get:
      security: []
      operationId: getSystemInfo
      responses:
        '200':
          description: Server and command-station capabilities.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemInfo' }
  /api/v1/auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/refresh:
    post:
      security: []
      operationId: refreshSession
      description: >-
        Rotates both tokens. The previous access and refresh tokens become invalid
        immediately. Existing WebSocket connections close no later than the
        original access-token expiry and must reconnect with the new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 20 }
      responses:
        '200':
          description: Rotated access and refresh tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/track-power:
    get:
      operationId: getTrackPowerStatus
      deprecated: true
      description: Compatibility alias for the complete command-station status.
      responses:
        '200':
          description: Current command-station status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StationStatus' }
    put:
      operationId: setTrackPower
      description: >-
        Enables or disables track power. Requires drive permission. Power-off is
        a priority safety command. A successful explicit power-on clears the
        emergency-stop interlock and permits new driving commands.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '204': { description: Track-power command accepted by the station driver. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/emergency-stop:
    post:
      operationId: emergencyStop
      description: >-
        Sends the command station's global priority emergency-stop command.
        Requires drive permission. Positive throttle and function commands remain
        inhibited until a subsequent explicit power-on succeeds.
      responses:
        '204': { description: Emergency-stop command accepted by the station driver. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/station/status:
    get:
      operationId: getStationStatus
      description: Queries the command station for its current power, operating, fault and telemetry status.
      responses:
        '200':
          description: Current command-station status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StationStatus' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses:
        '204': { description: Session revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/me:
    get:
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/locomotives:
    get:
      operationId: listLocomotives
      responses:
        '200':
          description: Rolling-stock locomotives.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Locomotive' }
    post:
      operationId: createLocomotive
      description: Creates a locomotive. Requires the administrator role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocomotiveInput' }
      responses:
        '201':
          description: Locomotive created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Locomotive' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/locomotives/{locomotiveId}:
    parameters:
      - $ref: '#/components/parameters/LocomotiveId'
    get:
      operationId: getLocomotive
      responses:
        '200':
          description: Locomotive details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Locomotive' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateLocomotive
      description: Replaces the writable locomotive properties. Requires the administrator role and no active control lease.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocomotiveInput' }
      responses:
        '200':
          description: Locomotive updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Locomotive' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: deleteLocomotive
      description: Deletes an unused locomotive. Requires the administrator role. Locomotives referenced by control history are retained and return 409.
      responses:
        '204': { description: Locomotive deleted. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/locomotives/{locomotiveId}/control-lease:
    parameters:
      - $ref: '#/components/parameters/LocomotiveId'
    post:
      operationId: acquireLocomotiveControl
      description: Atomically acquires exclusive control of a free locomotive.
      responses:
        '201':
          description: Exclusive control acquired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlLease' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Another live lease already owns the locomotive.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/control-leases/{leaseId}/heartbeat:
    parameters:
      - $ref: '#/components/parameters/LeaseId'
    put:
      operationId: heartbeatControlLease
      description: >-
        Renews the lease inactivity deadline. A heartbeat received at or after
        expiresAt is rejected and never revives an expired unswept lease.
      responses:
        '200':
          description: Lease renewed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlLease' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/control-leases/{leaseId}:
    parameters:
      - $ref: '#/components/parameters/LeaseId'
    delete:
      operationId: releaseControlLease
      description: Commands speed zero and moves the lease to stopping before release.
      responses:
        '202': { description: Controlled stop and release accepted. }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/locomotives/{locomotiveId}/throttle:
    parameters:
      - $ref: '#/components/parameters/LocomotiveId'
    put:
      operationId: setLocomotiveThrottle
      description: >-
        Sets speed and direction and renews the active lease inactivity deadline.
        Speed zero is a safety command and preempts queued ordinary commands. A
        positive speed is rejected while track power is off or unknown, while an
        emergency stop is active, or when a safety command preempted it while queued.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ThrottleCommand' }
      responses:
        '204': { description: Command accepted by the station driver. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/locomotives/{locomotiveId}/functions/{functionNumber}:
    parameters:
      - $ref: '#/components/parameters/LocomotiveId'
      - name: functionNumber
        in: path
        required: true
        schema: { type: integer, minimum: 0, maximum: 68 }
    put:
      operationId: setLocomotiveFunction
      description: >-
        Sets a locomotive function and renews the active lease inactivity deadline.
        The command is rejected while track power is off or unknown, while an
        emergency stop is active, or when a safety command preempted it while queued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leaseId, enabled]
              properties:
                leaseId: { type: string, format: uuid }
                enabled: { type: boolean }
      responses:
        '204': { description: Function command accepted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/blocks:
    get:
      operationId: listBlocks
      responses:
        '200':
          description: Blocks and their current occupancy.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Block' }
  /api/v1/turnouts:
    get:
      operationId: listTurnouts
      responses:
        '200':
          description: Turnouts.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Turnout' }
  /api/v1/turnouts/{turnoutId}:
    parameters:
      - name: turnoutId
        in: path
        required: true
        schema: { type: string }
    put:
      operationId: setTurnout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state: { type: string, enum: [straight, diverging] }
      responses:
        '204': { description: Turnout command accepted. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/routes:
    get:
      operationId: listRoutes
      responses:
        '200':
          description: Routes and their operational states.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Route' }
  /api/v1/routes/{routeId}/reserve:
    parameters: [{ $ref: '#/components/parameters/RouteId' }]
    post:
      operationId: reserveRoute
      responses:
        '204': { description: Route reserved atomically. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/routes/{routeId}/activate:
    parameters: [{ $ref: '#/components/parameters/RouteId' }]
    post:
      operationId: activateRoute
      responses:
        '204': { description: Required turnouts commanded and route activated. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/routes/{routeId}/release:
    parameters: [{ $ref: '#/components/parameters/RouteId' }]
    post:
      operationId: releaseRoute
      responses:
        '204': { description: Route released. }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/exports/rolling-stock:
    get:
      operationId: exportRollingStock
      description: Downloads a versioned ZIP archive containing the rolling-stock library.
      responses:
        '200':
          description: Rolling-stock archive.
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/vnd.dcc-control.package+zip:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/imports/rolling-stock:
    post:
      operationId: importRollingStock
      description: Imports a native rolling-stock archive. Requires the administrator role.
      parameters:
        - $ref: '#/components/parameters/ImportMode'
      requestBody:
        required: true
        content:
          application/vnd.dcc-control.package+zip:
            schema: { type: string, format: binary }
          application/zip:
            schema: { type: string, format: binary }
      responses:
        '204': { description: Archive imported. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/layout/export:
    get:
      operationId: exportLayout
      description: Downloads blocks, turnouts, routes and feedback mappings as a versioned ZIP archive.
      responses:
        '200':
          description: Layout archive.
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/vnd.dcc-control.package+zip:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/layout/import:
    post:
      operationId: importLayout
      description: Imports a native layout archive. Requires the administrator role.
      parameters:
        - $ref: '#/components/parameters/ImportMode'
      requestBody:
        required: true
        content:
          application/vnd.dcc-control.package+zip:
            schema: { type: string, format: binary }
          application/zip:
            schema: { type: string, format: binary }
      responses:
        '204': { description: Archive imported. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/events:
    get:
      operationId: openEventWebSocket
      description: >-
        Upgrades to WebSocket. The first message is a complete system.snapshot.
        Clients ignore events whose sequence is less than or equal to the latest
        snapshot sequence and request another snapshot after a gap. The connection
        closes when its access token expires or its session is revoked. Event
        schemas are defined in asyncapi.yaml.
      responses:
        '101': { description: WebSocket established. }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    LocomotiveId:
      name: locomotiveId
      in: path
      required: true
      schema: { type: string }
    LeaseId:
      name: leaseId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RouteId:
      name: routeId
      in: path
      required: true
      schema: { type: string }
    ImportMode:
      name: mode
      in: query
      required: false
      description: Merge archive objects into the database or replace the corresponding library.
      schema: { type: string, enum: [merge, replace], default: merge }
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing, invalid, expired or revoked access token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated role is not authorised.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: >-
        Operational state prevents the command. Safety-related problem codes are
        `emergency_stop_active`, `track_power_off`, `track_power_unknown` and
        `safety_command_preempted`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ServiceUnavailable:
      description: Command station is offline. The problem code is `station_offline`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, category]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        category:
          type: string
          enum: [authentication, authorization, validation, not_found, conflict, safety, station_unavailable, internal]
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password, clientId]
      properties:
        username: { type: string, minLength: 3, maxLength: 32 }
        password: { type: string, minLength: 10 }
        clientId: { type: string, minLength: 1, maxLength: 128 }
        clientName: { type: string, maxLength: 128 }
        platform: { type: string, maxLength: 64 }
    TokenPair:
      type: object
      required: [accessToken, refreshToken, accessExpiresAt, refreshExpiresAt, sessionId, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        accessExpiresAt: { type: string, format: date-time }
        refreshExpiresAt: { type: string, format: date-time }
        sessionId: { type: string, format: uuid }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      required: [id, username, role, enabled, mustChangePassword, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [viewer, driver, dispatcher, administrator] }
        enabled: { type: boolean }
        mustChangePassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time }
    Capabilities:
      type: object
      required: [driver, trackPower, locomotiveControl, functions, maxFunctionNumber, accessoryControl, feedback]
      properties:
        driver: { type: string, enum: [simulator, dccex, z21] }
        trackPower: { type: boolean }
        locomotiveControl: { type: boolean }
        functions:
          type: integer
          minimum: 0
          description: Number of supported function numbers, retained for compatibility.
        maxFunctionNumber:
          type: integer
          minimum: 0
          description: Inclusive maximum function number accepted by the current driver.
        accessoryControl: { type: boolean }
        feedback: { type: boolean }
    StationStatus:
      type: object
      additionalProperties: false
      required: [connectivity, trackPower, emergencyStop, shortCircuit, programmingMode, mainCurrentMilliAmps, programmingCurrentMilliAmps, filteredMainCurrentMilliAmps, temperatureCelsius, supplyVoltageMilliVolts, trackVoltageMilliVolts, highTemperature, powerLost, externalShortCircuit, internalShortCircuit]
      properties:
        connectivity: { type: string, enum: [online, degraded, offline] }
        lastSeen: { type: string, format: date-time }
        trackPower: { type: string, enum: [on, off, unknown] }
        emergencyStop: { type: boolean }
        shortCircuit: { type: boolean }
        programmingMode: { type: boolean }
        mainCurrentMilliAmps: { type: integer }
        programmingCurrentMilliAmps: { type: integer }
        filteredMainCurrentMilliAmps: { type: integer }
        temperatureCelsius: { type: integer }
        supplyVoltageMilliVolts: { type: integer, minimum: 0 }
        trackVoltageMilliVolts: { type: integer, minimum: 0 }
        highTemperature: { type: boolean }
        powerLost: { type: boolean }
        externalShortCircuit: { type: boolean }
        internalShortCircuit: { type: boolean }
    SystemInfo:
      type: object
      required: [serverVersion, apiVersion, minimumClientApiVersion, eventApiVersion, minimumClientEventApiVersion, station]
      properties:
        serverVersion: { type: string }
        apiVersion: { type: string }
        minimumClientApiVersion: { type: string }
        eventApiVersion: { type: string }
        minimumClientEventApiVersion: { type: string }
        station: { $ref: '#/components/schemas/Capabilities' }
    LocomotiveInput:
      type: object
      additionalProperties: false
      required: [name, dccAddress, addressKind, speedSteps]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        dccAddress: { type: integer, minimum: 1, maximum: 10239 }
        addressKind: { type: string, enum: [short, long] }
        speedSteps: { type: integer, enum: [14, 28, 128] }
        manufacturer: { type: string, maxLength: 128 }
        model: { type: string, maxLength: 128 }
    Locomotive:
      allOf:
        - $ref: '#/components/schemas/LocomotiveInput'
        - type: object
          required: [id]
          properties:
            id: { type: string }
    ControlLease:
      type: object
      required: [id, locomotiveId, userId, sessionId, state, acquiredAt, renewedAt, expiresAt, heartbeatMillis]
      properties:
        id: { type: string, format: uuid }
        locomotiveId: { type: string }
        userId: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        state: { type: string, enum: [active, stopping, released] }
        acquiredAt: { type: string, format: date-time }
        renewedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        releaseAfter: { type: string, format: date-time }
        releaseReason: { type: string }
        heartbeatMillis: { type: integer, minimum: 100 }
    ThrottleCommand:
      type: object
      additionalProperties: false
      required: [leaseId, speed, direction]
      properties:
        leaseId: { type: string, format: uuid }
        speed: { type: integer, minimum: 0, maximum: 100 }
        direction: { type: string, enum: [forward, reverse] }
    Block:
      type: object
      required: [id, name, occupied]
      properties:
        id: { type: string }
        name: { type: string }
        occupied: { type: boolean }
    Turnout:
      type: object
      required: [id, name, dccAddress, desiredState, reportedState]
      properties:
        id: { type: string }
        name: { type: string }
        dccAddress: { type: integer, minimum: 1 }
        desiredState: { type: string, enum: [straight, diverging] }
        reportedState: { type: string, enum: [straight, diverging, unknown] }
    Route:
      type: object
      required: [id, name, state]
      properties:
        id: { type: string }
        name: { type: string }
        state: { type: string, enum: [idle, reserved, active] }
        reservedBySession: { type: string }